Script to downsize Video files in a folder
I just got an FFmpeg.exe from online: https://ffmpeg.org/download.html
Created a folder and dropped the "Downsize Folder.bat" together with the "ffmpeg.exe" into it an run the Batch file.
It will ask what folder you want to downsize, and after any key, starts re-encoding file by file.

This script keeps all the original Video, Audio and Subtitle Tracks, just re-encodes it and lessens the quality. (this is the "CFG" parameter, 0 is best quality and 52 is worst)
Worked great to downsize the 1.7GB files of a large series .MKV files to 250/300MB a pop, without any visual loss of quality.
If applied to a Live Action show, you might want to lessen the CFG value a bit.